Mock/requirements.txt
mysql-connector-python==9.0.0
Faker==30.3.0
numpy==2.1.2
pillow==10.4.0